requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pymongo==4.7.2
dnspython==2.6.1
bcrypt==4.1.3
cryptography==42.0.7
python-dotenv==1.0.1
PyJWT==2.8.0
